Added converting fixed point wrapper around filter chain

+++ lib/inc/FixedPoint.h
@@ -8,14 +8,14 @@
 
 using namespace cnl::literals;
 
-template <int I, int F, class Narrowest>
+template <uint8_t I, uint8_t F, class Narrowest>
 using elastic_fixed_point = cnl::fixed_point<
     cnl::elastic_integer<
         I + F,
         Narrowest>,
     -F>;
 
-template <int I, int F, class Narrowest>
+template <uint8_t I, uint8_t F, class Narrowest>
 using safe_elastic_fixed_point = cnl::fixed_point<
     cnl::overflow_integer<
         cnl::elastic_integer<
